requirements.txt
Flask==1.1.1
Werkzeug==0.15.4
Flask_Cors==3.0.8
Pillow==7.2.0
gunicorn==20.0.4
